refactor: flow directory
