+++ src/qdash/api/routers/calibration.py
@@ -9,10 +9,6 @@ from prefect.client.orchestration import PrefectClient
 from pydantic import BaseModel
 from qdash.api.lib.auth import get_current_active_user
 from qdash.api.schemas.auth import User
-from qdash.api.schemas.calibration import (
-    ExecuteCalibRequest,
-    ExecuteCalibResponse,
-)
 from qdash.api.schemas.exception import InternalSeverError
 from qdash.config import Settings, get_settings
 from qdash.dbmodel.calibration_note import CalibrationNoteDocument
@@ -88,38 +84,7 @@ def get_calibration_note(
 
 
 
-@router.post(
-    "/calibration",
-    response_model=ExecuteCalibResponse,
-    summary="Executes a calibration by creating a flow run from a deployment.",
-    operation_id="execute_calib",
-)
-async def execute_calib(
-    request: ExecuteCalibRequest,
-    current_user: Annotated[User, Depends(get_current_active_user)],
-    settings: Annotated[Settings, Depends(get_settings)],
-) -> ExecuteCalibResponse:
-    """Create a flow run from a deployment."""
-    client = PrefectClient(api=settings.prefect_api_url)
-    logger.info(f"current user: {current_user.username}")
-    env = settings.env
-    target_deployment = await client.read_deployment_by_name(f"main/{env}-main")
-    try:
-        execution_id = generate_execution_id(current_user.username, request.chip_id)
-        TagDocument.insert_tags(request.tags, current_user.username)
-        resp = await client.create_flow_run_from_deployment(
-            deployment_id=target_deployment.id,
-            parameters={"menu": request.model_dump(), "execution_id": execution_id},
-        )
-    except Exception as e:
-        logger.warning(e)
-        raise InternalSeverError(detail=f"Failed to execute calibration {e!s}")
-    logger.warning(resp)
-    chip_id = ChipDocument.get_current_chip(current_user.username).chip_id
-    return ExecuteCalibResponse(
-        flow_run_url=f"http://127.0.0.1:{settings.prefect_port}/flow-runs/flow-run/{resp.id}",
-        qdash_ui_url=f"http://127.0.0.1:{settings.ui_port}/execution/{chip_id}/{execution_id}",
-    )
+
 
 
 ja = dateutil.tz.gettz("Asia/Tokyo")
