+++ src/qdash/workflow/engine/session/qubex.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 from qdash.dbmodel.calibration_note import CalibrationNoteDocument
 from qdash.workflow.engine.session.base import BaseSession
-from qdash.workflow.utils.merge_notes import merge_notes_by_timestamp
+from qdash.workflow._internal.merge_notes import merge_notes_by_timestamp
 
 
 class QubexSession(BaseSession):
